pyproject.toml
[tool.poetry]
name = "hw13"
version = "0.1.0"
description = ""
authors = ["data-tamer2410 <vitaliy.dayneko2410@gmail.com>"]
packages = [{'include' = 'fashion_mnist_classifier_2_0'}]

[tool.poetry.dependencies]
python = "3.11.*"
tensorflow = "^2.18.0"
streamlit = "^1.40.2"
numpy = "<2.1.0"
pandas = "^2.2.3"
pillow = "^11.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
